refactoring the way biota species is handled now using WoRMS as a single point of truth

+++ marisco/nc_template.py
@@ -32,7 +32,7 @@ class NCTemplater:
         self.dim = cdl['dim']
         self.enum_types = {}
 
-# %% ../nbs/api/nc_template.ipynb 8
+# %% ../nbs/api/nc_template.ipynb 9
 @patch
 def nuclide_vars(
     self:NCTemplater,
@@ -60,7 +60,7 @@ def nuclide_vars(
         )
     ]
 
-# %% ../nbs/api/nc_template.ipynb 11
+# %% ../nbs/api/nc_template.ipynb 12
 @patch
 def derive(
     self:NCTemplater,
@@ -74,14 +74,14 @@ def derive(
         'attrs': {key: nuclide['attrs'][key] + suffix['attrs'][key] for key in nuclide['attrs']}
         }
 
-# %% ../nbs/api/nc_template.ipynb 17
+# %% ../nbs/api/nc_template.ipynb 18
 @patch
 def create_enum_types(self:NCTemplater):
     "Create enumeration types"
     for name, enum in self.enum_dicts.items(): 
-        self.enum_types[name] = self.nc.createEnumType(np.uint16, name, enum)
+        self.enum_types[name] = self.nc.createEnumType(np.uint, name, enum)
 
-# %% ../nbs/api/nc_template.ipynb 18
+# %% ../nbs/api/nc_template.ipynb 19
 @patch
 def create_groups(self:NCTemplater):
     "Create NetCDF groups"
@@ -90,7 +90,7 @@ def create_groups(self:NCTemplater):
         grp = self.nc.createGroup(grp_name)
         self.create_variables(grp)
 
-# %% ../nbs/api/nc_template.ipynb 19
+# %% ../nbs/api/nc_template.ipynb 20
 @patch
 def create_variables(self:NCTemplater, 
                      grp:netCDF4.Group, # NetCDF group
@@ -101,7 +101,7 @@ def create_variables(self:NCTemplater,
         self.create_group_specific_variables(grp)
         self.create_analyte_variables(grp)
 
-# %% ../nbs/api/nc_template.ipynb 20
+# %% ../nbs/api/nc_template.ipynb 21
 @patch
 def create_default_variables(self:NCTemplater, 
                              grp:netCDF4.Group, # NetCDF group
@@ -110,7 +110,7 @@ def create_default_variables(self:NCTemplater,
         vars = self.cdl['vars']['defaults'].values()
         for var in vars: self.create_variable(grp, var)
 
-# %% ../nbs/api/nc_template.ipynb 21
+# %% ../nbs/api/nc_template.ipynb 22
 @patch
 def create_group_specific_variables(self:NCTemplater, 
                              grp:netCDF4.Group, # NetCDF group
@@ -120,7 +120,7 @@ def create_group_specific_variables(self:NCTemplater,
         for var in vars.get(name2grp(grp.name, self.cdl), {}).values(): 
             self.create_variable(grp, var)
 
-# %% ../nbs/api/nc_template.ipynb 22
+# %% ../nbs/api/nc_template.ipynb 23
 @patch
 def create_analyte_variables(self:NCTemplater, 
                              grp:netCDF4.Group, # NetCDF group
@@ -131,7 +131,7 @@ def create_analyte_variables(self:NCTemplater,
         for v in self.cdl['vars']['suffixes'].values(): 
             self.create_variable(grp, self.derive(var, v))
 
-# %% ../nbs/api/nc_template.ipynb 23
+# %% ../nbs/api/nc_template.ipynb 24
 @patch
 def create_variable(self:NCTemplater, 
                     grp:netCDF4.Group, # NetCDF group
@@ -144,10 +144,11 @@ def create_variable(self:NCTemplater,
                                 self.dim['name'])
     nc_var.setncatts(attrs) 
 
-# %% ../nbs/api/nc_template.ipynb 24
+# %% ../nbs/api/nc_template.ipynb 26
 @patch
 def generate(self:NCTemplater):
     "Generate CDL"
+    # with NetCDFWriter(self.tpl_fname) as self.nc:
     with Dataset(self.tpl_fname, 'w', format='NETCDF4') as self.nc:
         self.nc.setncatts(self.cdl['global_attrs']) 
         self.create_enum_types()
